+++ internal/module/address/domain/address_repository.go
@@ -7,5 +7,5 @@ type AddressRepository interface {
 	Create(ctx context.Context, address *Address) error
 	Update(ctx context.Context, address Address) error
 	// FindByCEP(ctx context.Context, cep string) (Address, error)
-	FindByIDCompany(ctx context.Context, idCompany uint) (*Address, error)
+	FindByIDCompany(ctx context.Context, idCompany int64) (*Address, error)
 }
