+++ internal/module/company/application/dto/create_company_request.go
@@ -12,7 +12,7 @@ type CreateCompanyRequest struct {
 	Name         string                                 `json:"name" validate:"required|min=3"`
 	TradeName    string                                 `json:"trade_name" validate:"required|min=3"`
 	CNPJ         string                                 `json:"cnpj" validate:"required|len=18|numeric"`
-	RegisteredBy uint                                   `json:"registered_by" validate:"required|numeric"`
+	RegisteredBy int64                                  `json:"registered_by" validate:"required|numeric"`
 	Address      addr.CreateAddressRequest              `json:"address,omitzero"`
 	Phone        phone.CreatePhoneRequest               `json:"phone,omitzero"`
 	Email        email.CreateEmailRequest               `json:"email,omitzero"`
