+++ internal/module/company/application/dto/update_company_request.go
@@ -9,7 +9,7 @@ import (
 )
 
 type UpdateCompanyRequest struct {
-	IDCompany   uint                                   `json:"id_company" validate:"required|numeric"`
+	IDCompany   int64                                  `json:"id_company" validate:"required|numeric"`
 	Name        string                                 `json:"name" validate:"required|min=3"`
 	TradeName   string                                 `json:"trade_name" validate:"required|min=3"`
 	CNPJ        string                                 `json:"cnpj" validate:"required|len=18|numeric"`
