+++ internal/module/company/domain/company_address_repository.go
@@ -4,6 +4,6 @@ package domain
 import "context"
 
 type CompanyAddressRepository interface {
-	Create(ctx context.Context, idCompany, idAddress uint) error
-	DeleteByIDCompany(ctx context.Context, idCompany uint) error
+	Create(ctx context.Context, idCompany, idAddress int64) error
+	DeleteByIDCompany(ctx context.Context, idCompany int64) error
 }
