+++ internal/module/company/domain/company_email_repository.go
@@ -4,6 +4,6 @@ package domain
 import "context"
 
 type CompanyEmailRepository interface {
-	Create(ctx context.Context, idCompany, idEmail uint) error
-	DeleteByIDCompany(ctx context.Context, idCompany uint) error
+	Create(ctx context.Context, idCompany, idEmail int64) error
+	DeleteByIDCompany(ctx context.Context, idCompany int64) error
 }
