+++ internal/module/company/domain/company_phone_repository.go
@@ -4,6 +4,6 @@ package domain
 import "context"
 
 type CompanyPhoneRepository interface {
-	Create(ctx context.Context, idCompany, idPhone uint) error
-	DeleteByIDCompany(ctx context.Context, idCompany uint) error
+	Create(ctx context.Context, idCompany, idPhone int64) error
+	DeleteByIDCompany(ctx context.Context, idCompany int64) error
 }
