+++ internal/module/company/infrastructure/persistence/database/postgres/company_email_postgres.go
@@ -18,14 +18,14 @@ func NewCompanyEmail(db postgres.Executor) *CompanyEmail {
 	}
 }
 
-func (c CompanyEmail) Create(ctx context.Context, idCompany, idEmail uint) error {
+func (c CompanyEmail) Create(ctx context.Context, idCompany, idEmail int64) error {
 	query := `INSERT INTO company_email (id_companies, id_emails, started_at)
 		VALUES ($1, $2, $3)`
 	_, err := c.db.Exec(ctx, query, idCompany, idEmail, time.Now())
 	return err
 }
 
-func (c CompanyEmail) DeleteByIDCompany(ctx context.Context, idCompany uint) error {
+func (c CompanyEmail) DeleteByIDCompany(ctx context.Context, idCompany int64) error {
 	query := `UPDATE company_email SET ended_at = $1 WHERE id_companies = $2 AND ended_at IS NULL`
 	_, err := c.db.Exec(ctx, query, time.Now(), idCompany)
 	return err
