+++ internal/module/company/infrastructure/persistence/database/postgres/company_phone_postgres.go
@@ -18,14 +18,14 @@ func NewCompanyPhone(db postgres.Executor) *CompanyPhone {
 	}
 }
 
-func (c CompanyPhone) Create(ctx context.Context, idCompany, idPhone uint) error {
+func (c CompanyPhone) Create(ctx context.Context, idCompany, idPhone int64) error {
 	query := `INSERT INTO company_phone (id_companies, id_phones, started_at)
 		VALUES ($1, $2, $3)`
 	_, err := c.db.Exec(ctx, query, idCompany, idPhone, time.Now())
 	return err
 }
 
-func (c CompanyPhone) DeleteByIDCompany(ctx context.Context, idCompany uint) error {
+func (c CompanyPhone) DeleteByIDCompany(ctx context.Context, idCompany int64) error {
 	query := `UPDATE company_phone SET ended_at = $1 WHERE id_companies = $2 AND ended_at IS NULL`
 	_, err := c.db.Exec(ctx, query, time.Now(), idCompany)
 	return err
