+++ internal/module/email/application/dto/update_email_request.go
@@ -2,6 +2,6 @@
 package dto
 
 type UpdateEmailRequest struct {
-	ID      uint   `json:"id_email" validate:"required"`
+	ID      int64  `json:"id_email" validate:"required"`
 	Address string `json:"address" validate:"required|email"`
 }
