+++ internal/module/email/domain/email_repository.go
@@ -9,5 +9,5 @@ type EmailRepository interface {
 	Create(ctx context.Context, email *Email) error
 	Update(ctx context.Context, email Email) error
 	Delete(ctx context.Context, email Email) error
-	FindByIDCompany(ctx context.Context, idCompany uint) (*Email, error)
+	FindByIDCompany(ctx context.Context, idCompany int64) (*Email, error)
 }
