+++ internal/module/phone/application/dto/update_phone_request.go
@@ -2,7 +2,7 @@
 package dto
 
 type UpdatePhoneRequest struct {
-	ID         uint   `json:"id_phone" validate:"required|numeric"`
+	ID         int64  `json:"id_phone" validate:"required|numeric"`
 	Number     string `json:"number" validate:"required|min=13|max=14"`
 	Kind       string `json:"kind" validate:"required"`
 	Department string `json:"department,omitzero"`
