+++ internal/module/phone/domain/phone_repository.go
@@ -9,5 +9,5 @@ type PhoneRepository interface {
 	Create(ctx context.Context, phone *Phone) error
 	Update(ctx context.Context, phone Phone) error
 	Delete(ctx context.Context, phone Phone) error
-	FindByIDCompany(ctx context.Context, idCompany uint) (*Phone, error)
+	FindByIDCompany(ctx context.Context, idCompany int64) (*Phone, error)
 }
