+++ internal/module/social_media/application/dto/social_media_read_model.go
@@ -4,8 +4,8 @@ package dto
 import "time"
 
 type SocialMediaReadModel struct {
-	ID        uint      `json:"id"`
-	IDCompany uint      `json:"id_company"`
+	ID        int64     `json:"id"`
+	IDCompany int64     `json:"id_company"`
 	Platform  string    `json:"platform"`
 	URL       string    `json:"url"`
 	CreatedAt time.Time `json:"created_at"`
