+++ internal/module/social_media/application/dto/update_social_media_request.go
@@ -2,7 +2,7 @@
 package dto
 
 type UpdateSocialMediaRequest struct {
-	IDSocialMedia uint   `json:"id_social_media" validate:"required|numeric"`
+	IDSocialMedia int64  `json:"id_social_media" validate:"required|numeric"`
 	Platform      string `json:"platform" validate:"required"`
 	URL           string `json:"url" validate:"required|url"`
 }
