+++ internal/module/social_media/domain/social_media_repository.go
@@ -9,6 +9,6 @@ type SocialMediaRepository interface {
 	Create(ctx context.Context, socialMedia *SocialMedia) error
 	Update(ctx context.Context, socialMedia SocialMedia) error
 	Delete(ctx context.Context, socialMedia SocialMedia) error
-	FindByIDCompany(ctx context.Context, idCompany uint) ([]*SocialMedia, error)
-	DeleteByIDCompany(ctx context.Context, idCompany uint) error
+	FindByIDCompany(ctx context.Context, idCompany int64) ([]*SocialMedia, error)
+	DeleteByIDCompany(ctx context.Context, idCompany int64) error
 }
